use shared cattr converter
